tests/wpt/web-platform-tests/web-bundle/subresource-loading/script-element-removal.https.tentative.html

<!DOCTYPE html>
<title>On-going subresource loading should fail immediately when the web bundle element is removed</title>
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>
<script>
  window.TEST_WEB_BUNDLE_ELEMENT_TYPE = 'script';
  setup(() => {
    assert_true(HTMLScriptElement.supports('webbundle'));
  });
</script>
<body>
  <script src="resources/element-removal-test.js"></script>
</body>
